fix store user and restructure profile
