+++ src/components/Profile/ProfileTitle/ProfileTitle.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../globalStyle/mixins';
+@import '../../../globalStyle/mixins';
 
 .profileTitle {
   @include Header2;
