--- src/components/UserInfo/UserInfo.module.scss
@@ -1,130 +1,0 @@
-@use '../../globalStyle/mixins.scss' as m;
-@use '../../globalStyle/variables' as v;
-.UserInfo {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    max-width: 410px;
-    width: 100%;
-    min-height: 677px;
-    height: 1px;
-    margin-bottom: 32px;
-}
-.avatar {
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-
-    &_block {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-    &_plug {
-        width: 100px;
-        height: 100px;
-        border-radius: 50%;
-        background-color: v.$white-click;
-    }
-    &_userNick {
-        @include m.Body1;
-        margin-top: 32px;
-    }
-    &_datalist {
-        list-style-type: none;
-        margin: 0;
-        padding: 0;
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-        align-items: center;
-        flex-wrap: wrap;
-        margin-top: 80px;
-    }
-    &_list {
-        @include m.Body2;
-        color: v.$text-green-color;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        max-width: 347px;
-        width: 100%;
-        margin-bottom: 12px;
-    }
-    &_list:last-of-type {
-        margin-bottom: 0;
-        padding-bottom: 68px;
-        border-bottom: 1px solid rgba(75, 75, 75 , .25);
-    }
-    &_span {
-        @include m.Body2;
-        color: v.$text-green-color;
-        margin-top: 6px;
-    }
-    &_set {
-        color: transparent;
-        width: 100%;
-        height: 47px;
-        margin-top: 40px;
-        margin-bottom: 24px;
-        padding: 0;
-    }
-    &_set::-webkit-file-upload-button {
-        visibility: hidden;
-    }
-    &_set::before {
-        @include m.Link;
-        content: 'Загрузить фото';
-        color: v.$white-color;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background: v.$accent-color;
-        border-radius: 8px;
-        outline: none;
-        white-space: nowrap;
-        cursor: pointer;
-        width: 100%;
-        height: 100%;
-    }
-    &_set:hover::before {
-        background: #007873;
-    }
-}
-.Links {
-    display: flex;
-    flex-direction: column;
-    margin-top: 32px;
-    align-items: flex-start;
-    max-width: 347px;
-    width: 100%;
-    
-    &_block {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        height: 100%;
-        justify-content: space-between;
-    }
-    &_description {
-        @include m.Body2;
-        margin-bottom: 12px;
-    }
-    &_item {
-        @include m.Body2;
-        margin-bottom: 12px;
-    }
-    &_item:hover {
-        opacity: 0.7;
-      }
-    &_item:last-of-type {
-        margin-bottom: 0;
-    }
-}
-.ProfilleButton {
-    @include m.Body2;
-}
